refactor: move deployment loader impls to core
