+++ packages/core/package.json
@@ -92,6 +92,7 @@
     "@ethersproject/address": "5.6.1",
     "debug": "^4.3.2",
     "ethers": "^5.4.7",
+    "fs-extra": "^10.0.0",
     "js-graph-algorithms": "1.0.18",
     "lodash": "4.17.21",
     "ndjson": "2.0.0",
